test(p2p_task): handle incoming proposal part proptest
