+++ crates/pathfinder/src/consensus/inner.rs
@@ -8,9 +8,6 @@ mod integration_testing;
 mod p2p_task;
 mod persist_proposals;
 
-#[cfg(all(test, feature = "p2p"))]
-mod p2p_task_tests;
-
 #[cfg(test)]
 mod test_helpers;
